rename contourgeometry module
